requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==3.0.4
channels-redis==3.4.1
constantly==15.1.0
cryptography==39.0.2
daphne==3.0.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.5
packaging==23.0
pycountry==22.3.5
parameterized==0.9.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pytest==7.2.2
pytz==2022.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
zope.event==4.5.0
zope.interface==5.5.2